fix: landing opinion
